refactor models to one file
