+++ index.ts
@@ -1,3 +1,2 @@
 export { ShContextMenuModule } from './src/sh-context-menu.module';
-export { IShContextMenuItem } from './src/sh-context-item';
-export { IShContextOptions } from './src/sh-context-options';
+export { IShContextMenuItem, IShContextOptions } from './src/sh-context-menu.models';
