--- src/sh-context-default-options.ts
@@ -1,6 +1,0 @@
-import { IShContextOptions } from './sh-context-options';
-
-export const ShContextDefaultOptions: IShContextOptions = {
-  rtl: false,
-  theme:'light'
-};
