--- src/sh-context-options.ts
@@ -1,6 +1,0 @@
-﻿//pr test
-
-export interface IShContextOptions {
-  rtl?: boolean;
-  theme?: 'light' | 'dark'
-};
